Coldfusion NeeviaPDF integration
This code is designed to integrate NeeviaPDF compression tool with coldfusion 9 and above. Note: ***At the time of this entry 6/12/2015, CF 9 & 10 are not supporting COM object integration. Alternatively Cfexecute can be used in its place. ***

This code integrates with the cffileupload module and checks a table called tbl_file_uploads for the last files uploaded using a dateTimeStamp = Now() variable. The value of dateTimeStamp is inserted into the tbl_file_uploads table as column fileDate.
We now want to compress the PDF files just uploaded. 

We run a query with a where clause set to where fileDate >= dateTimeStamp AND fileupload Like '% .pdf' . This restricts the results returned to only those files in that date time stamp range and who's name ends with .pdf extension. 

Once we have that we set a loop from the query and loop through the query. delcaring a new variable in the loop, its set to each value returned from the query and the loop runs the cfexecute block for each result returned, finding the file in the path declared and compressing it then returning the compressed file with the original name.